1318 - move beneficiary folder to admin
