+++ src/components/admin/beneficiary/CreateForm.tsx
@@ -1,38 +1,30 @@
-import React from 'react'
-import { useMutation, useQueryClient, UseQueryResult } from '@tanstack/react-query'
+import React, { useState } from 'react'
+import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { useRouter } from 'next/router'
 import { useTranslation } from 'next-i18next'
 
 import { AxiosError, AxiosResponse } from 'axios'
 import * as yup from 'yup'
-import {
-  Box,
-  Button,
-  FormControl,
-  Grid,
-  InputLabel,
-  MenuItem,
-  Select,
-  Typography,
-} from '@mui/material'
+import { Box, Button, Grid, Typography } from '@mui/material'
 
 import { BeneficiaryFormData, BeneficiaryListResponse } from 'gql/beneficiary'
 import { BeneficiaryType } from './BeneficiaryTypes'
 import { routes } from 'common/routes'
 import { ApiErrors } from 'service/apiErrors'
-import { useViewBeneficiary, useEditBeneficiary } from 'service/beneficiary'
+import { useCreateBeneficiary } from 'service/beneficiary'
 import { useCompaniesList } from 'service/company'
 import { AlertStore } from 'stores/AlertStore'
-import PersonSelect from 'components/person/PersonSelect'
+import PersonSelect from 'components/admin/person/PersonSelect'
 import GenericForm from 'components/common/form/GenericForm'
 import FormTextField from 'components/common/form/FormTextField'
 import SubmitButton from 'components/common/form/SubmitButton'
-import CitySelect from 'components/cities/CitySelect'
-import CountrySelect from 'components/countries/CountrySelect'
+import CitySelect from 'components/admin/cities/CitySelect'
+import CountrySelect from 'components/admin/countries/CountrySelect'
 import { endpoints } from 'service/apiEndpoints'
-import CompanySelect from 'components/companies/CompanySelect'
+import CompanySelect from 'components/admin/companies/CompanySelect'
 import Link from 'components/common/Link'
 import OrganizerRelationSelect from './OrganizerRelationSelect'
+import BeneficiaryTypeSelect from './BeneficiaryTypeSelect'
 
 const validationSchema = yup
   .object()
@@ -62,28 +54,27 @@ const validationSchema = yup
     }),
   })
 
-export default function EditForm() {
+const initialValues: BeneficiaryFormData = {
+  type: BeneficiaryType.individual,
+  personId: '',
+  companyId: '',
+  countryCode: 'BG',
+  cityId: '',
+  description: '',
+  publicData: '',
+  privateData: '',
+  organizerRelation: 'none',
+  campaigns: [],
+}
+
+export default function CreateForm() {
   const queryClient = useQueryClient()
   const router = useRouter()
   const { t } = useTranslation()
-  const id = router.query.id
   const { data: companies } = useCompaniesList()
-  const { data }: UseQueryResult<BeneficiaryListResponse> = useViewBeneficiary(String(id))
+  const [beneficiaryType, setBeneficiaryType] = useState<string>(BeneficiaryType.individual)
 
-  const initialValues = {
-    type: data?.type,
-    cityId: data?.cityId || '',
-    companyId: data?.companyId || undefined,
-    countryCode: data?.countryCode || '',
-    description: data?.description || '',
-    personId: data?.personId || undefined,
-    privateData: data?.privateData || '',
-    publicData: data?.publicData || '',
-    organizerRelation: data?.organizerRelation || '',
-    campaigns: data?.campaigns || [],
-  }
-
-  const mutationFn = useEditBeneficiary(String(id))
+  const mutationFn = useCreateBeneficiary()
   const mutation = useMutation<
     AxiosResponse<BeneficiaryListResponse>,
     AxiosError<ApiErrors>,
@@ -101,8 +92,8 @@ export default function EditForm() {
   async function onSubmit(values: BeneficiaryFormData) {
     const data = {
       type: values.type,
-      personId: values.personId,
-      companyId: values.companyId,
+      personId: values.type == BeneficiaryType.individual ? values.personId : undefined,
+      companyId: values.type == BeneficiaryType.company ? values.companyId : undefined,
       countryCode:
         companies?.find((c) => c.id === values.companyId)?.countryCode || values.countryCode,
       cityId: companies?.find((c) => c.id === values.companyId)?.cityId || values.cityId,
@@ -120,33 +111,14 @@ export default function EditForm() {
       validationSchema={validationSchema}>
       <Box sx={{ height: '62.6vh', marginBottom: '9%' }}>
         <Typography variant="h5" component="h2" sx={{ textAlign: 'center' }}>
-          {t('beneficiary:forms:edit-heading')}
+          {t('beneficiary:forms:add-heading')}
         </Typography>
         <Typography variant="body1" sx={{ textAlign: 'center', padding: '8px', mb: 2 }}>
           {t('beneficiary:forms:info')}
         </Typography>
         <Grid container spacing={3}>
           <Grid item xs={6}>
-            <FormControl fullWidth variant="outlined" size="small">
-              <InputLabel>{t('beneficiary:grid:type')}</InputLabel>
-              <Select
-                fullWidth
-                name="type"
-                disabled={Boolean(id)}
-                defaultValue={initialValues.type}
-                label={t('beneficiary:grid:type')}>
-                <MenuItem value={undefined} disabled>
-                  <em>None</em>
-                </MenuItem>
-                {Object.values(BeneficiaryType)?.map((type) => {
-                  return (
-                    <MenuItem key={type} value={type}>
-                      {t('beneficiary:grid:' + type)}
-                    </MenuItem>
-                  )
-                })}
-              </Select>
-            </FormControl>
+            <BeneficiaryTypeSelect name={'type'} setBeneficiaryType={setBeneficiaryType} />
           </Grid>
           <Grid item xs={6}>
             <OrganizerRelationSelect
@@ -154,12 +126,12 @@ export default function EditForm() {
               label="beneficiary:grid:organizerRelation"
             />
           </Grid>
-          {initialValues.type === BeneficiaryType.individual ? (
+          {beneficiaryType === BeneficiaryType.individual ? (
             <>
               <Grid item xs={12}>
                 <Typography paddingLeft={'inherit'} marginBottom={2}>
                   {t('beneficiary:forms.labels.person-select')}{' '}
-                  <Link href={routes.admin.person.create}>
+                  <Link href={routes.admin.person.create} target="_blank">
                     {t('beneficiary:forms.labels.create-new')}
                   </Link>
                 </Typography>
@@ -176,7 +148,7 @@ export default function EditForm() {
             <Grid item xs={12}>
               <Typography paddingLeft={'inherit'} marginBottom={2}>
                 {t('beneficiary:forms.labels.company-select')}{' '}
-                <Link href={routes.admin.company.create}>
+                <Link href={routes.admin.company.create} target="_blank">
                   {t('beneficiary:forms.labels.create-new')}
                 </Link>
               </Typography>
